+++ src/command/webview/gif/CatgirlGifCommand.ts
@@ -2,16 +2,16 @@ import Nekoslife from '../../../api/nekoslife/Nekoslife'
 import PictureCommand from '../../PictureCommand'
 import { window } from 'vscode'
 
-export default class LewdCatgirl extends PictureCommand {
+export default class CatgirlGif extends PictureCommand {
   public constructor() {
-    super('lewd-catgirl', { nsfw: true })
+    super('catgirl-gif')
   }
 
   public async run(): Promise<void> {
     const API = new (this.getNekoslife())
 
     if (API instanceof Nekoslife.v2) {
-      await API.fetchNsfwBody('lewd')
+      await API.fetchSfwBody('ngif')
         .then((body) => {
           if (body.msg) window.showErrorMessage(body.msg)
           else this.createWebviewPanel(body.url)
@@ -22,13 +22,7 @@ export default class LewdCatgirl extends PictureCommand {
     }
 
     if (API instanceof Nekoslife.v3) {
-      const array = [
-        API.fetchNsfwImg('neko_lewd'),
-        API.fetchNsfwImg('neko_ero')
-      ]
-      const fetchImg = array[Math.floor(Math.random() * array.length)]
-
-      await fetchImg
+      await API.fetchSfwGif('neko')
         .then((body) => {
           if (body.data.status.code !== 200) {
             window.showErrorMessage(`HTTPS CODE: ${body.data.status.code} : ${body.data.status.message ?? 'No Message'}`)
