+++ src/command/webview/picture/CatgirlCommand.ts
@@ -2,16 +2,16 @@ import Nekoslife from '../../../api/nekoslife/Nekoslife'
 import PictureCommand from '../../PictureCommand'
 import { window } from 'vscode'
 
-export default class LewdFoxgirl extends PictureCommand {
+export default class Catgirl extends PictureCommand {
   public constructor() {
-    super('lewd-foxgirl', { nsfw: true })
+    super('catgirl')
   }
 
   public async run(): Promise<void> {
     const API = new (this.getNekoslife())
 
     if (API instanceof Nekoslife.v2) {
-      await API.fetchNsfwBody('lewdk')
+      await API.fetchSfwBody('neko')
         .then((body) => {
           if (body.msg) window.showErrorMessage(body.msg)
           else this.createWebviewPanel(body.url)
@@ -22,13 +22,7 @@ export default class LewdFoxgirl extends PictureCommand {
     }
 
     if (API instanceof Nekoslife.v3) {
-      const array = [
-        API.fetchNsfwImg('kitsune_lewd'),
-        API.fetchNsfwImg('kitsune_ero')
-      ]
-      const fetchImg = array[Math.floor(Math.random() * array.length)]
-
-      await fetchImg
+      await API.fetchSfwImg('neko')
         .then((body) => {
           if (body.data.status.code !== 200) {
             window.showErrorMessage(`HTTPS CODE: ${body.data.status.code} : ${body.data.status.message ?? 'No Message'}`)
