+++ src/command/webview/picture/FoxgirlCommand.ts
@@ -1,17 +1,17 @@
-import Nekoslife from '../../api/nekoslife/Nekoslife'
-import PictureCommand from '../PictureCommand'
+import Nekoslife from '../../../api/nekoslife/Nekoslife'
+import PictureCommand from '../../PictureCommand'
 import { window } from 'vscode'
 
-export default class CatgirlGif extends PictureCommand {
+export default class Foxgirl extends PictureCommand {
   public constructor() {
-    super('catgirl-gif')
+    super('foxgirl')
   }
 
   public async run(): Promise<void> {
     const API = new (this.getNekoslife())
 
     if (API instanceof Nekoslife.v2) {
-      await API.fetchSfwBody('ngif')
+      await API.fetchSfwBody('fox_girl')
         .then((body) => {
           if (body.msg) window.showErrorMessage(body.msg)
           else this.createWebviewPanel(body.url)
@@ -22,7 +22,7 @@ export default class CatgirlGif extends PictureCommand {
     }
 
     if (API instanceof Nekoslife.v3) {
-      await API.fetchSfwGif('neko')
+      await API.fetchSfwImg('kitsune')
         .then((body) => {
           if (body.data.status.code !== 200) {
             window.showErrorMessage(`HTTPS CODE: ${body.data.status.code} : ${body.data.status.message ?? 'No Message'}`)
