+++ src/command/webview/picture/nsfw/LewdCatgirlCommand.ts
@@ -1,17 +1,17 @@
-import Nekoslife from '../../api/nekoslife/Nekoslife'
-import PictureCommand from '../PictureCommand'
+import Nekoslife from '../../../../api/nekoslife/Nekoslife'
+import PictureCommand from '../../../PictureCommand'
 import { window } from 'vscode'
 
-export default class Foxgirl extends PictureCommand {
+export default class LewdCatgirl extends PictureCommand {
   public constructor() {
-    super('foxgirl')
+    super('lewd-catgirl', { nsfw: true })
   }
 
   public async run(): Promise<void> {
     const API = new (this.getNekoslife())
 
     if (API instanceof Nekoslife.v2) {
-      await API.fetchSfwBody('fox_girl')
+      await API.fetchNsfwBody('lewd')
         .then((body) => {
           if (body.msg) window.showErrorMessage(body.msg)
           else this.createWebviewPanel(body.url)
@@ -22,7 +22,13 @@ export default class Foxgirl extends PictureCommand {
     }
 
     if (API instanceof Nekoslife.v3) {
-      await API.fetchSfwImg('kitsune')
+      const array = [
+        API.fetchNsfwImg('neko_lewd'),
+        API.fetchNsfwImg('neko_ero')
+      ]
+      const fetchImg = array[Math.floor(Math.random() * array.length)]
+
+      await fetchImg
         .then((body) => {
           if (body.data.status.code !== 200) {
             window.showErrorMessage(`HTTPS CODE: ${body.data.status.code} : ${body.data.status.message ?? 'No Message'}`)
