+++ src/command/webview/picture/nsfw/LewdFoxgirlCommand.ts
@@ -1,17 +1,17 @@
-import Nekoslife from '../../api/nekoslife/Nekoslife'
-import PictureCommand from '../PictureCommand'
+import Nekoslife from '../../../../api/nekoslife/Nekoslife'
+import PictureCommand from '../../../PictureCommand'
 import { window } from 'vscode'
 
-export default class Catgirl extends PictureCommand {
+export default class LewdFoxgirl extends PictureCommand {
   public constructor() {
-    super('catgirl')
+    super('lewd-foxgirl', { nsfw: true })
   }
 
   public async run(): Promise<void> {
     const API = new (this.getNekoslife())
 
     if (API instanceof Nekoslife.v2) {
-      await API.fetchSfwBody('neko')
+      await API.fetchNsfwBody('lewdk')
         .then((body) => {
           if (body.msg) window.showErrorMessage(body.msg)
           else this.createWebviewPanel(body.url)
@@ -22,7 +22,13 @@ export default class Catgirl extends PictureCommand {
     }
 
     if (API instanceof Nekoslife.v3) {
-      await API.fetchSfwImg('neko')
+      const array = [
+        API.fetchNsfwImg('kitsune_lewd'),
+        API.fetchNsfwImg('kitsune_ero')
+      ]
+      const fetchImg = array[Math.floor(Math.random() * array.length)]
+
+      await fetchImg
         .then((body) => {
           if (body.data.status.code !== 200) {
             window.showErrorMessage(`HTTPS CODE: ${body.data.status.code} : ${body.data.status.message ?? 'No Message'}`)
